+++ backend/canisters/user/api/src/updates/send_message_v2.rs
@@ -30,7 +30,6 @@ pub enum Response {
     InvalidRequest(String),
     TransferFailed(String),
     TransferCannotBeZero,
-    TransferLimitExceeded(u128),
     UserSuspended,
     InternalError(String),
 }