+++ frontend/openchat-agent/src/services/user/candid/idl.d.ts
@@ -3,10 +3,8 @@ import {
     _SERVICE,
     ReplyContext,
     ChatSummaryUpdates,
-    GroupChatSummaryUpdates,
     DirectChatSummaryUpdates,
     ChatEventWrapper,
-    CreateGroupArgs,
     CreateGroupResponse,
     DeleteGroupResponse,
     InitialStateV2Response,
@@ -30,6 +28,8 @@ import {
     ProposalContent,
     PrizeContent,
     PrizeWinnerContent,
+    MessageReminderCreated,
+    MessageReminder,
     Proposal,
     PollContent,
     PollVotes,
@@ -96,6 +96,7 @@ import {
     Icrc1Account,
     DeletedMessageResponse,
     GroupGate,
+    SetMessageReminderResponse,
 } from "./types";
 export {
     _SERVICE as UserService,
@@ -193,6 +194,9 @@ export {
     DeletedMessageResponse as ApiDeletedDirectMessageResponse,
     SendMessageWithTransferToGroupResponse as ApiSendMessageWithTransferToGroupResponse,
     GroupGate as ApiGroupGate,
+    SetMessageReminderResponse as ApiSetMessageReminderResponse,
+    MessageReminder as ApiMessageReminder,
+    MessageReminderCreated as ApiMessageReminderCreated,
 };
 
 export const idlFactory: IDL.InterfaceFactory;